requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.10.1
aiofiles==23.2.1
requests==2.31.0
